+++ canaille/oidc/endpoints/tokens.py
@@ -2,6 +2,7 @@ import datetime
 
 from flask import Blueprint
 from flask import abort
+from flask import current_app
 from flask import flash
 from flask import request
 
@@ -42,6 +43,10 @@ def view(user, token):
         elif request.form.get("action") == "revoke":
             token.revokation_date = datetime.datetime.now(datetime.timezone.utc)
             Backend.instance.save(token)
+            request_ip = request.remote_addr or "unknown IP"
+            current_app.logger.security(
+                f"Revoked token for {token.subject.user_name} in client {token.client.client_name} by {user.user_name} from {request_ip}"
+            )
             flash(_("The token has successfully been revoked."), "success")
 
         else: